api_gateway/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bson==0.5.10
certifi==2024.8.30
click==8.1.7
confluent-kafka==2.6.1
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
iniconfig==2.0.0
motor==3.6.0
packaging==24.2
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.10.1
pydantic_core==2.27.1
pymongo==4.9.2
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-multipart==0.0.19
redis==5.2.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
